Dob Club python code
Data class assignment based on the dog's
eligibility to stay in the Bing's club.
To be eligibl, the dog needs to be in the breed
that I wanted.

At first it will print out the information of the dog,
then the eligiblity on the dog to be in Bing's club.